requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
asyncer==0.0.8
bleach==6.2.0
bokeh==3.7.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.2.0
funkify==0.4.5
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.35.0
ipywidgets==8.1.6
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jupyter_bokeh==4.0.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
linkify-it-py==2.0.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
narwhals==1.35.0
nest-asyncio==1.6.0
numpy==2.2.5
ollama==0.4.8
openai==1.76.0
packaging==25.0
pandas==2.2.3
panel==1.6.2
param==2.2.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyviz_comms==3.0.4
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.4.0
watchfiles==1.0.5
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.14
xtyping==0.9.0
xyzservices==2025.1.0
